Print specific worksheets in Excel to PDF
So my girlfriend works in a nursing home and has all this printing and stuff to do every month by a deadline and she's super stressed about it. 

I'm all like, "Babe, why?" 
And she's all like, "I dunno..."
And so I'm all like, "Ima make you a spreadsheet."
And she's all like, "Cool."

Weeks later, she's all stressed again saying, "I have to right-click all these worksheets to get them to save as individual PDF's and there are like 117 of them and I have to do it every single month and it sucks."

And so I'm all like, "Babe, why?"
And she's like, "I dunno."
And so I'm all like, "Babe, I'ma make you a script to automate that shit."
And then she's all like, "Cool."

So I found this awesome script on StackOverflow (such a sweet web site) and modified it to be able to do a bunch of worksheets in one go. You have to install pypiwin32 to get it to work on Windows, but hey it's a small thing isn't it?